Pass input with all dimensions to basis.
