+++ skfda/_utils/__init__.py
@@ -3,4 +3,5 @@ from . import constants
 from ._utils import (_list_of_arrays, _cartesian_product,
                      _check_estimator, parameter_aliases,
                      _to_grid, check_is_univariate,
-                     _same_domain, _to_array_maybe_ragged)
+                     _same_domain, _to_array_maybe_ragged,
+                     _reshape_eval_points)
