+++ tests/test_basis_evaluation.py
@@ -1,5 +1,6 @@
 
-from skfda.representation.basis import FDataBasis, Monomial, BSpline, Fourier
+from skfda.representation.basis import (
+    FDataBasis, Monomial, BSpline, Fourier, Constant, VectorValued)
 import unittest
 
 import numpy as np
@@ -412,6 +413,43 @@ class TestBasisEvaluationMonomial(unittest.TestCase):
 
             np.testing.assert_array_almost_equal(f(t).round(3), res)
             np.testing.assert_array_almost_equal(f.evaluate(t).round(3), res)
+
+
+class TestBasisEvaluationVectorValued(unittest.TestCase):
+
+    def test_vector_valued_constant(self):
+
+        basis_first = Constant()
+        basis_second = Constant()
+
+        basis = VectorValued([basis_first, basis_second])
+
+        fd = FDataBasis(basis=basis, coefficients=[[1, 2], [3, 4]])
+
+        self.assertEqual(fd.dim_codomain, 2)
+
+        res = np.array([[[1, 2]], [[3, 4]]])
+
+        np.testing.assert_allclose(fd(0), res)
+
+    def test_vector_valued_constant_monomial(self):
+
+        basis_first = Constant(domain_range=(0, 5))
+        basis_second = Monomial(n_basis=3, domain_range=(0, 5))
+
+        basis = VectorValued([basis_first, basis_second])
+
+        fd = FDataBasis(basis=basis, coefficients=[
+                        [1, 2, 3, 4], [3, 4, 5, 6]])
+
+        self.assertEqual(fd.dim_codomain, 2)
+
+        np.testing.assert_allclose(fd.domain_range[0], (0, 5))
+
+        res = np.array([[[1, 2], [1, 9], [1, 24]],
+                        [[3, 4], [3, 15], [3, 38]]])
+
+        np.testing.assert_allclose(fd([0, 1, 2]), res)
 
 
 if __name__ == '__main__':
